fixed images
This fixes the aspect ratio for the images in the blog.  You will no longer need to resize images for the blog expect if they are really, really big and take up too much space. ( Note: Remember to resize/crop the profile images on the about page to make sure they don't get stretched. )